Updated integ-tests implementation for OSD

Changes updated


add -o parameter


Updated test cases
