+++ src/test_workflow/integ_test/integ_test_suite_opensearch.py
@@ -6,23 +6,29 @@
 # compatible open source license.
 
 import glob
+import json
 import logging
 import os
 from pathlib import Path
 from typing import Any
 
+from git.git_repository import GitRepository
 from manifests.build_manifest import BuildManifest
 from manifests.bundle_manifest import BundleManifest
+from paths.script_finder import ScriptFinder
+from system.execute import execute
 from test_workflow.dependency_installer_opensearch import DependencyInstallerOpenSearch
 from test_workflow.integ_test.integ_test_suite import IntegTestSuite, InvalidTestConfigError
-from test_workflow.integ_test.topology import Topology
+from test_workflow.integ_test.topology import NodeEndpoint, Topology
 from test_workflow.test_recorder.test_recorder import TestRecorder
+from test_workflow.test_recorder.test_result_data import TestResultData
 from test_workflow.test_result.test_component_results import TestComponentResults
 from test_workflow.test_result.test_result import TestResult
 
 
 class IntegTestSuiteOpenSearch(IntegTestSuite):
     dependency_installer: DependencyInstallerOpenSearch
+    repo: GitRepository
 
     def __init__(
         self,
@@ -42,6 +48,12 @@ class IntegTestSuiteOpenSearch(IntegTestSuite):
             dependency_installer_opensearch,
             bundle_manifest_opensearch,
             build_manifest_opensearch
+        )
+        self.repo = GitRepository(
+            self.component.repository,
+            self.component.commit_id,
+            os.path.join(self.work_dir, self.component.name),
+            test_config.working_directory
         )
 
     def execute_tests(self) -> TestComponentResults:
@@ -88,6 +100,43 @@ class IntegTestSuiteOpenSearch(IntegTestSuite):
             os.chdir(self.work_dir)
             self.pretty_print_message("Running integration tests for " + self.component.name + " " + config)
             return self.multi_execute_integtest_sh(endpoints, security, config)
+
+    def multi_execute_integtest_sh(self, cluster_endpoints: list, security: bool, test_config: str) -> int:
+        script = ScriptFinder.find_integ_test_script(self.component.name, self.repo.working_directory)
+
+        def custom_node_endpoint_encoder(node_endpoint: NodeEndpoint) -> dict:
+            return {"endpoint": node_endpoint.endpoint, "port": node_endpoint.port, "transport": node_endpoint.transport}
+        if os.path.exists(script):
+            if len(cluster_endpoints) == 1:
+                single_data_node = cluster_endpoints[0].data_nodes[0]
+                cmd = f"bash {script} -b {single_data_node.endpoint} -p {single_data_node.port} -s {str(security).lower()} -v {self.bundle_manifest.build.version}"
+            else:
+                endpoints_list = []
+                for cluster_details in cluster_endpoints:
+                    endpoints_list.append(cluster_details.__dict__)
+                endpoints_string = json.dumps(endpoints_list, indent=0, default=custom_node_endpoint_encoder)
+                cmd = f"bash {script} -e '"
+                cmd = cmd + endpoints_string + "'"
+                cmd = cmd + f" -s {str(security).lower()} -v {self.bundle_manifest.build.version}"
+            self.repo_work_dir = os.path.join(
+                self.repo.dir, self.test_config.working_directory) if self.test_config.working_directory is not None else self.repo.dir
+            (status, stdout, stderr) = execute(cmd, self.repo_work_dir, True, False)
+            test_result_data = TestResultData(
+                self.component.name,
+                test_config,
+                status,
+                stdout,
+                stderr,
+                self.test_artifact_files
+            )
+            self.save_logs.save_test_result_data(test_result_data)
+            if stderr:
+                logging.info("Stderr reported for component: " + self.component.name)
+                logging.info(stderr)
+            return status
+        else:
+            logging.info(f"{script} does not exist. Skipping integ tests for {self.component.name}")
+            return 0
 
     @property
     def test_artifact_files(self) -> dict:
